+++ patch.go
@@ -123,8 +123,12 @@ var UnitTypes = map[string]([]UnitParameter){
 		{Name: "stereo", MinValue: 0, MaxValue: 1, CanSet: true, CanModulate: false},
 		{Name: "attack", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true, DisplayFunc: compressorTimeDispFunc},
 		{Name: "release", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true, DisplayFunc: compressorTimeDispFunc},
-		{Name: "invgain", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true},
-		{Name: "threshold", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true},
+		{Name: "invgain", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true, DisplayFunc: func(v int) (string, string) {
+			return strconv.FormatFloat(20*math.Log10(128/float64(v)), 'f', 2, 64), "dB"
+		}},
+		{Name: "threshold", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true, DisplayFunc: func(v int) (string, string) {
+			return strconv.FormatFloat(20*math.Log10(float64(v)/128), 'f', 2, 64), "dB"
+		}},
 		{Name: "ratio", MinValue: 0, MaxValue: 128, CanSet: true, CanModulate: true, DisplayFunc: func(v int) (string, string) { return formatFloat(1 - float64(v)/128), "" }}},
 	"speed": []UnitParameter{},
 	"out": []UnitParameter{